build: Moving test files to dedicated test directory and omitting from build.
